renaming the files and deleted un-necessary or duplicate files
