+++ src/screens/Detail/Detail.css
